% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdcov.R
\docType{package}
\name{pfica-package}
\alias{pfica}
\alias{pfica-package}
\title{Functional independent component analysis}
\description{
\code{pfica} Performs penalized (and non-penalized) independent component analysis for univariate functional data. Two alternative versions are implemented, both based on the spectral decomposition of the kurtosis operator. Our methods are iterfaced with the basis functions provided in the \strong{fda} package.
}
\details{
This package contains a set of tools for performing functional independent component analysis (FICA) by analysing the kurtosis structure of whitened data. Two FICA versions are considered depending on the basis representation of the data: \code{\link{ffobi}} computes the independent components using a canonical basis of functions and \code{\link{kffobi}} uses the weight functions of the covariance operator. The latter is also extended using a discrete penalty in  \code{\link{pspline.kffobi}}, being this function computationally more efficient. The current FICA routines use Mahalanobis kernel whitening and shrinkage covariance estimators to improve the outcomes in the estimation process.}
\references{
Ramsay, J. and B. Silverman (2005). \emph{Functional Data Analysis}. Springer.

Schafer, J. and K. Strimmer (2005). A shrinkage approach to large-scale covariance matrix estimation and implications for functional genomics. Statistical Applications in Genetics and Molecular Biology, 4.32(1).

Vidal, M., M. Rosso  and AM. Aguilera (2021). Smoothed independent component analysis of EEG data. \emph{Submitted}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-vidal/pfica}
}
}
\author{
Marc Vidal \email{marc.vidalbadia@ugent.be},
Ana Mª Aguilera \email{aaguiler@ugr.es}
}
